requirements.txt
APScheduler==3.9.1
click==8.0.4
configparser==5.2.0
<<<<<<< HEAD
=======
PyYAML>=6.0
>>>>>>> feature/configurable-concurrency
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
numpy==1.21.5
oura==1.1.4
pandas==1.3.5
PyMySQL==1.0.2
pyocclient==0.6
scikit-learn==1.0.2
SQLAlchemy==1.4.32
stravalib==0.10.4
sweat==0.4.0
tekore==4.3.0
Werkzeug==2.0.3
withings-api==2.4.0
greenlet<0.5,>=0.4.5